카테고리 베스트 아이템 버그 수정 및 카테고리 아이템 쿼리 성능 최적화

+++ src/main/java/org/example/tamaapi/dto/responseDto/CustomPageable.java
@@ -26,7 +26,7 @@ public class CustomPageable {
     }
 
     //직접 만든 페이징. 페이지 1부터 시작
-    public CustomPageable(int page, int size, int rowCount) {
+    public CustomPageable(int page, int size, Long rowCount) {
         this.page = page;
         this.size = size;
         this.pageCount = (int) Math.ceil((double) rowCount/size);
